+++ frontend/pages/list.php
@@ -1,0 +1,13 @@
+<!-- -------------------------------------------------
+
+   rmWebUI - Web interface for the reMarkable(R) cloud.
+
+   (c) 2021-2022 Fabien Pollet <polletfa@posteo.de>
+   MIT License (see LICENSE file)
+  
+  --------------------------------------------------- -->
+<div class="row">
+  <table class="table table-hover" id="files-table">
+  </table>
+</div>
+<script><?php require_once __DIR__."/list.js" ?></script>